Add test to read a real quince.properties file
