+++ WebApp/junit/resources/configuration/quince.properties
@@ -1,0 +1,17 @@
+app.urlstub=http://localhost:8080/QuinCe
+email.starttls=true
+email.ssl=false
+email.hostname=
+email.port=
+email.username=
+email.password=
+email.fromname=QuinCe
+email.fromaddress=quince@uib.no
+filestore=/Users/zuj007/Documents/QuinCe/VOYAGE_FILE_STORE
+extract_routines.configfile=./WebApp/junit/resources/configuration/extract_routines_config.csv
+qc_routines.configfile=./WebApp/junit/resources/configuration/qc_routines_config.csv
+columns.configfile=./WebApp/junit/resources/configuration/eqpco2_column_config.csv
+export.configfile=./WebApp/junit/resources/configuration/export_config.csv
+sensors.configfile=./WebApp/junit/resources/configuration/sensor_config.csv
+runtypes.configfile=./WebApp/junit/resources/configuration/run_types_config.csv
+map.max_points=1000
